bump oidc-spring-support to 0.2.11 and enable filter for token expiration * refactoring due to api changes * new property for expiration threshold, i.e. when to return a response header stating token needs refreshing
