+++ pom.xml
@@ -25,7 +25,7 @@
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
         <springfox.version>2.8.0</springfox.version>
-        <oidc-spring-support.version>0.2.3</oidc-spring-support.version>
+        <oidc-spring-support.version>0.2.11</oidc-spring-support.version>
     </properties>
     <dependencies>
         <dependency>
@@ -140,9 +140,15 @@
         </dependency>
         <dependency>
             <groupId>no.nav.security</groupId>
-            <artifactId>oidc-spring-test</artifactId>
+            <artifactId>oidc-test-support</artifactId>
             <version>${oidc-spring-support.version}</version>
             <scope>test</scope>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.springframework.boot</groupId>
+                    <artifactId>spring-boot-starter-jersey</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
         <dependency>
             <groupId>org.testcontainers</groupId>
