+++ src/main/java/no/nav/foreldrepenger/selvbetjening/ApiApplication.java
@@ -1,14 +1,13 @@
 package no.nav.foreldrepenger.selvbetjening;
 
+import no.nav.security.spring.oidc.api.EnableOIDCTokenValidation;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
 import io.prometheus.client.hotspot.DefaultExports;
-import no.nav.security.spring.oidc.validation.api.EnableOIDCTokenValidation;
 
 @SpringBootApplication
-@ComponentScan
 @EnableOIDCTokenValidation(ignore = { "org.springframework", "springfox.documentation" })
 public class ApiApplication {
