+++ src/test/java/no/nav/foreldrepenger/selvbetjening/ApplicationLocal.java
@@ -2,6 +2,8 @@ package no.nav.foreldrepenger.selvbetjening;
 
 import static org.springframework.context.annotation.FilterType.ASSIGNABLE_TYPE;
 
+import no.nav.security.oidc.test.support.spring.TokenGeneratorConfiguration;
+import no.nav.security.spring.oidc.api.EnableOIDCTokenValidation;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
@@ -10,8 +12,7 @@ import org.springframework.context.annotation.Import;
 
 import io.prometheus.client.hotspot.DefaultExports;
 import no.nav.foreldrepenger.selvbetjening.felles.util.EnvUtil;
-import no.nav.security.spring.oidc.test.TokenGeneratorConfiguration;
-import no.nav.security.spring.oidc.validation.api.EnableOIDCTokenValidation;
+
 
 @SpringBootApplication
 @EnableOIDCTokenValidation(ignore = { "org.springframework", "springfox.documentation" })
