+++ src/test/resources/application.properties
@@ -26,3 +26,5 @@ stub.mottak=true
 http.proxy=
 
 TOGGLES_ENABLE_FORELDREPENGESOKNAD=true
+
+logging.level.no.nav: DEBUG
